functions_nested_loops

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,13 +8,13 @@
 int main(void)
 {
 	int counter = 0;
-	int i = 1;
+	long i = 1;
 	long f = 1;
 	long s = 1;
 
 	while (counter < 50)
 	{
-		printf("%li, ", i);
+		printf("%d, ", i);
 		f = s;
 		s = i;
 		i = s + f;
